+++ src/Twilio/Rest/Domain.cs
@@ -35,6 +35,7 @@ namespace Twilio.Rest
         public static readonly Domain Intelligence = new Domain("intelligence");
         public static readonly Domain IpMessaging = new Domain("ip-messaging");
         public static readonly Domain Lookups = new Domain("lookups");
+        public static readonly Domain Marketplace = new Domain("marketplace");
         public static readonly Domain Messaging = new Domain("messaging");
         public static readonly Domain Microvisor = new Domain("microvisor");
         public static readonly Domain Monitor = new Domain("monitor");