+++ CHANGES.md
@@ -1,6 +1,21 @@
 twilio-csharp Changelog
 =======================
 
+[2024-06-06] Version 7.1.1
+--------------------------
+**Library - Fix**
+- [PR #748](https://github.com/twilio/twilio-csharp/pull/748): Adding TaskRouterGrant file. Thanks to [@tiwarishubham635](https://github.com/tiwarishubham635)!
+
+**Api**
+- Mark MaxPrice as obsolete
+
+**Lookups**
+- Update examples for `phone_number_quality_score`
+
+**Messaging**
+- List tollfree verifications on parent account and all sub-accounts
+
+
 [2024-05-24] Version 7.1.0
 --------------------------
 **Library - Docs**